Stick to double for correlation times

+++ Sources/Stats/mc_stats.hpp
@@ -28,7 +28,7 @@ struct Stats {
   /// Average of in-chain variances of the observable over all Markov Chains.
   double variance;
   /// Autocorrelation time
-  int correlation;
+  double correlation;
   /// Convergence estimator. The closer it is to 1, the better has the sampling
   /// converged.
   double R;
